checkpoint my work. Completed things to change, working on things to add

+++ src/main/java/rocks/zipcode/atm/bank/Account.java
@@ -32,7 +32,7 @@ public abstract class Account {
         return getBalance() >= amount;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return accountData.getBalance();
     }
